==> firstIteration/3-preprocesamiento.ipynb <==
import pandas as pd
from sklearn.preprocessing import MinMaxScaler
from sklearn.preprocessing import MultiLabelBinarizer

df = pd.read_parquet("../data/df_step1.parquet")

# Preprocesamiento

### Datos Faltantes

# Nulos
null_counts = df.isnull().sum()
print("Cantidad de valores nulos por columna (solo las que tienen nulos):", null_counts[null_counts > 0])

# Reemplazar valores nulos en 'Time signature' con la moda
mode_time_sig = df['Time signature'].mode()[0]
df['Time signature'].fillna(mode_time_sig, inplace=True)

#### Duplicados

# Analizar duplicados
duplicates = df.duplicated().sum()
df.drop_duplicates(inplace=True)
print(f"Cantidad de duplicados eliminados: {duplicates}")

#### Lyrics con menos de 100 caracteres

# Analizar columna Lyrics con largo de texto menor a 100 caracteres
lyrics_length = df['Lyrics'].str.len()
short_lyrics_count = (lyrics_length < 100).sum()
print(f"Cantidad de letras con menos de 10 caracteres: {short_lyrics_count}")

### Tratamiento de Multietiqueta

# Separar por comas y quitar espacios innecesarios
df['Genre_list'] = df['Genre'].str.split(',')

# Remover espacios al principio/final de cada etiqueta
df['Genre_list'] = df['Genre_list'].apply(lambda genres: [g.strip().lower() for g in genres])

# Show unique genres
unique_genres = set()
for genres in df['Genre_list']:
    unique_genres.update(genres)
print("Géneros únicos:", unique_genres)
print("Cantidad de géneros únicos:", len(unique_genres))


mlb = MultiLabelBinarizer()
genre_dummies = pd.DataFrame(mlb.fit_transform(df['Genre_list']),
                             columns=mlb.classes_,
                             index=df.index)

# Concatenar al dataframe original
df = pd.concat([df.drop(columns=['Genre', 'Genre_list']), genre_dummies], axis=1)

genre_counts = df[mlb.classes_].sum().sort_values(ascending=False)
genre_counts

### Variables categóricas a numéricas

# One-Hot Encoding
df = pd.get_dummies(df, columns=['Key', 'Time signature'], prefix=['Key', 'TS'], dtype=int)

### Normalización

# All columns except Lyrics and emotion
num_cols = df.columns.difference(['Lyrics', 'emotion'])
num_cols

scaler  = MinMaxScaler()
df[num_cols] = scaler .fit_transform(df[num_cols])

df.head()

# Guardar el DataFrame modificado en formato eficiente
df.to_parquet("../data/df_step2.parquet")

En la proxima iteración de preprocesamiento se hará el embedding de las letras de las canciones utilizando sentence-transformers.
