==> firstIteration/4-preprocesamiento.ipynb <==
import pandas as pd
from sentence_transformers import SentenceTransformer
from sklearn.decomposition import PCA
from sklearn.preprocessing import MinMaxScaler

df = pd.read_parquet("../data/df_step2.parquet")

df.info(verbose=True)

# Preprocesamiento 2

### Embeddings

# use a sample of the dataset, or the full dataset
useASample = True
if (useASample):
    df = df.sample(n=1000, random_state=42)

# Instanciás el modelo preentrenado
model = SentenceTransformer('all-MiniLM-L6-v2')

# Transformás las letras en embeddings
embeddings = model.encode(df['Lyrics'].tolist(), convert_to_tensor=False, show_progress_bar=True)

# Convertís los embeddings a DataFrame
embeddings_df = pd.DataFrame(embeddings, columns=[f'lyric_emb_{i}' for i in range(embeddings.shape[1])])

# Reseteás el índice de ambos DataFrames
df = df.reset_index(drop=True)
embeddings_df = embeddings_df.reset_index(drop=True)

# Concatenás al dataset original (sin la columna Lyrics)
df_final = pd.concat([df.drop(columns=['Lyrics']), embeddings_df], axis=1)

# Mostrás el resultado
df_final.head()

### PCA sobre embeddings

# 1. Inicializamos el PCA para retener el 80% de la varianza
pca = PCA(n_components=0.80, random_state=42)

# 2. Ajustamos y transformamos
lyrics_embeddings_pca = pca.fit_transform(embeddings_df)

# 3. Revisamos cuántos componentes quedaron
print(f"Componentes seleccionados: {pca.n_components_}")


### Normalizar los embeddings PCA

# Normalizamos los embeddings PCA
scaler = MinMaxScaler()

lyrics_embeddings_pca_normalized = scaler.fit_transform(lyrics_embeddings_pca)

# Convertimos a DataFrame
lyrics_embeddings_pca_normalized_df = pd.DataFrame(lyrics_embeddings_pca_normalized,
                                                    columns=[f'lyric_emb_pca_{i}' for i in range(lyrics_embeddings_pca_normalized.shape[1])])

# Concatenamos al dataset original
df_final_pca = pd.concat([df.drop(columns=['Lyrics']), lyrics_embeddings_pca_normalized_df], axis=1)

# Guardar el DataFrame modificado en formato eficiente
df_final_pca.to_parquet("../data/df_step3.parquet")

En el proximo paso se comenzará con los entrenamiento de los modelos RF y KNN
