==> firstIteration/5-entrenamiento-rf.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import classification_report, accuracy_score
from sklearn.model_selection import GridSearchCV

# df = pd.read_parquet("../data/df_step3.parquet")
df = pd.read_parquet("../data/df_step3.full.parquet")

df.info(verbose=True)

# Entrenamiento

X = df.drop(columns=["emotion"])
y = df["emotion"]

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, stratify=y, random_state=42
)

# ONLY FOR COLAB
# param_grid = {
#     'n_estimators': [100, 200],           # cantidad de árboles
#     'max_depth': [10, 20, None],          # profundidad máxima
#     'min_samples_split': [2, 5],          # mínimo de muestras para dividir un nodo
#     'min_samples_leaf': [1, 2],           # mínimo de muestras en una hoja
#     'max_features': ['sqrt', 'log2'],     # cantidad de features a usar en cada split
#     'bootstrap': [True, False]            # usar bootstrap samples o no
# }

# rf = RandomForestClassifier(random_state=42)

# grid_search = GridSearchCV(
#     estimator=rf,
#     param_grid=param_grid,
#     cv=5,
#     scoring='accuracy',
#     n_jobs=-1,
#     verbose=1
# )

# grid_search.fit(X_train, y_train)

# ONLY FOR COLAB
# print("Mejores parámetros:")
# print(grid_search.best_params_)

# print("Accuracy en test:")
# print(grid_search.score(X_test, y_test))

# rf = RandomForestClassifier(
#     n_estimators=grid_search.best_params_['n_estimators'],
#     max_depth=grid_search.best_params_['max_depth'],
#     min_samples_split=grid_search.best_params_['min_samples_split'],
#     min_samples_leaf=grid_search.best_params_['min_samples_leaf'],
#     max_features=grid_search.best_params_['max_features'],
#     bootstrap=grid_search.best_params_['bootstrap'],
#     random_state=42
# )
rf = RandomForestClassifier(
    n_estimators=200,
    max_depth=20,
    min_samples_split=2,
    min_samples_leaf=1,
    max_features='sqrt',
    bootstrap=False,
    random_state=42,
    class_weight='balanced',
    n_jobs=-1
)
rf.fit(X_train, y_train)

y_pred = rf.predict(X_test)
print("Accuracy:", accuracy_score(y_test, y_pred))
print("\nReporte de clasificación:")
print(classification_report(y_test, y_pred))
