==> firstIteration/6-entrenamiento-knn.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.neighbors import KNeighborsClassifier
from sklearn.metrics import accuracy_score, classification_report
from sklearn.model_selection import GridSearchCV

# df = pd.read_parquet("../data/df_step3.parquet")
df = pd.read_parquet("../data/df_step3.full.parquet")

df.info(verbose=True)

# Entrenamiento

X = df.drop(columns=["emotion"])
y = df["emotion"]

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, stratify=y, random_state=42
)

# ONLY FOR COLAB
# param_grid = {
#     'n_neighbors': [3, 5, 7, 9, 11],
#     'weights': ['uniform', 'distance'],
#     'metric': ['euclidean', 'manhattan'],
#     'leaf_size': [20, 30, 40],
# }

# knn = KNeighborsClassifier()

# grid_search = GridSearchCV(
#     estimator=knn,
#     param_grid=param_grid,
#     cv=3,
#     scoring='accuracy',  # se puede omitir, pero es buena práctica especificarlo
#     n_jobs=-1,
#     verbose=1
# )
# grid_search.fit(X_train, y_train)

# ONLY FOR COLAB
# print("Mejores parámetros encontrados:")
# print(grid_search.best_params_)

# print("Accuracy en test:")
# print(grid_search.score(X_test, y_test))

# knn = KNeighborsClassifier(
#     n_neighbors=grid_search.best_params_['n_neighbors'],
#     weights=grid_search.best_params_['weights'],
#     metric=grid_search.best_params_['metric'],
#     leaf_size=grid_search.best_params_['leaf_size']
# )
knn = KNeighborsClassifier(
    n_neighbors=11,
    weights='distance',
    metric='manhattan',
    leaf_size=20,
    n_jobs=-1
)
knn.fit(X_train, y_train)

y_pred = knn.predict(X_test)
print("Accuracy:", accuracy_score(y_test, y_pred))
print("\nReporte de clasificación:")
print(classification_report(y_test, y_pred))
